# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoRestTeach API
  description: |
    A teaching REST API built with **Go + Gin + GORM + PostgreSQL**.

    Designed to help Flutter teams learn how to work with REST endpoints and Postman.

    ## Authentication

    The API uses **JWT Bearer** authentication:
    1. Register via `POST /auth/register`
    2. Obtain tokens via `POST /auth/login`
    3. Add the header: `Authorization: Bearer <access_token>`
    4. The access token expires in **15 minutes**. Refresh it via `POST /auth/refresh`

    ## Response Format

    **Success:**
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "page": 1, "per_page": 10, "total": 42 }
    }
    ```

    **Error:**
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Validation failed",
        "details": [{ "field": "email", "message": "Must be a valid email address" }]
      }
    }
    ```

    ## Error Codes

    | HTTP | Code | When |
    |------|------|------|
    | 400 | `VALIDATION_ERROR` | Invalid request body fields |
    | 400 | `BAD_REQUEST` | Malformed JSON or missing body |
    | 401 | `UNAUTHORIZED` | Missing or invalid token |
    | 401 | `TOKEN_EXPIRED` | JWT access token has expired |
    | 403 | `FORBIDDEN` | Action not allowed (e.g. editing another user's post) |
    | 404 | `NOT_FOUND` | Resource does not exist |
    | 409 | `CONFLICT` | Email already registered |
...
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Paste the `access_token` received from `/auth/login`.
        Format: `Bearer <token>`
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              enum:
                - VALIDATION_ERROR
                - BAD_REQUEST
                - UNAUTHORIZED
                - TOKEN_EXPIRED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - FILE_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - INTERNAL_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        bio:
          type: string
          example: Flutter developer at AcidSoft
        avatar_id:
          type: string
          format: uuid
          nullable: true
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Use `GET /images/{avatar_id}` to fetch the image bytes
        created_at:
          type: string
          format: date-time
    TokenPair:
      type: object
      properties:
        access_token:
          type: string
          description: Short-lived JWT (15 minutes). Use in the Authorization header.
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        refresh_token:
          type: string
          description: Long-lived opaque UUID token (7 days). Store it securely.
          example: a3f4c2d1-1234-5678-abcd-ef0123456789
        token_type:
          type: string
          example: Bearer
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authorization header is required
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Post not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Must be a valid email address
                - field: password
                  message: Must be at least 8 characters long
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register
      description: |
        Creates a new user account.

        **Team task (beginner):** Register a new user and verify that trying to
        register with the same email a second time returns `409 CONFLICT`.
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: John Doe
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  minLength: 8
                  example: secret123
      responses:
        '201':
          description: User successfully created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/UserPublic'
              example:
                success: true
                data:
                  id: 550e8400-e29b-41d4-a716-446655440000
                  name: John Doe
                  email: john@example.com
                  bio: ""
                  avatar_id: null
                  created_at: "2026-02-21T13:00:00Z"
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email already registered
          content:
            application/json:
              example:
                success: false
                error:
                  code: CONFLICT
                  message: Email is already registered
  /auth/login:
    post:
      tags: [auth]
      summary: Login
      description: |
        Authenticates the user and returns an **access token** (JWT, 15 min)
        and a **refresh token** (UUID, 7 days).

        **Team task (beginner):** Log in and save both tokens. Use the
        `access_token` in the `Authorization` header for protected endpoints.
        Paste the token at [jwt.io](https://jwt.io) to inspect the claims.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: john@example.com
                password:
                  type: string
                  example: secret123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenPair'
              example:
                success: true
                data:
                  access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.abc123
                  refresh_token: a3f4c2d1-1234-5678-abcd-ef0123456789
                  token_type: Bearer
        '401':
          description: Invalid credentials
          content:
            application/json:
              example:
                success: false
                error:
                  code: UNAUTHORIZED
                  message: Invalid email or password
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: |
        Exchanges the refresh token for a **new access + refresh token pair**
        (rotation pattern). The old refresh token is deleted after a successful
        exchange — do not reuse it.

        **Team task (intermediate):** Wait for the access token to expire
        (or temporarily lower `JWT_ACCESS_EXPIRES_MINUTES` in `.env`), then
        call this endpoint to get a fresh token pair without logging in again.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: a3f4c2d1-1234-5678-abcd-ef0123456789
      responses:
        '200':
          description: New token pair issued
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenPair'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout
      description: |
        Invalidates the given refresh token by deleting it from the database.

        > **Note:** The access token **cannot be revoked** — it remains valid
        > until it expires (15 min). This is by design in stateless JWT
        > architectures. Keep the expiry short in production.

        **Team task (intermediate):** Log out, then try to use the old
        `refresh_token` again — you should receive `401 UNAUTHORIZED`.
      operationId: logout
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  example: a3f4c2d1-1234-5678-abcd-ef0123456789
      responses:
        '204':
          description: Logged out successfully (no body)
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── USERS ──────────────────────────────────────────────────────────────────
...
  /images/{id}:
    get:
      tags: [images]
      summary: Get image
      description: |
        Returns raw image bytes with the appropriate `Content-Type` header
        (`image/jpeg`, `image/png`, etc.).

        This endpoint is **public** — no authentication required.
        Use the UUID from `avatar_id` (user profile) or `image_id` (post).

        **In Flutter:**
        ```dart
        Image.network('http://localhost:8080/api/v1/images/$imageId')
        ```

        **In browser:** paste the full URL — the image renders directly.
      operationId: getImage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: 660e8400-e29b-41d4-a716-446655440001
      responses:
        '200':
          description: Raw image binary data
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
            image/gif:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFound'